requirements.txt
biopython
PyYAML
openai
tiktoken
requests
nltk